Refactor to pull camera controller out of sandbox

+++ src/main/java/uk/ashleybye/avalon/Application.java
@@ -51,7 +51,7 @@ public abstract class Application {
     while (running) {
 
       timer.tick();
-      float dt = (float) timer.getDeltaSeconds();
+      double dt = timer.getDeltaSeconds();
 
       for (var layer : layers) {
         layer.onUpdate(dt);
